Allow to get array shape as a TensorShape
