+++ api/src/main/kotlin/org/jetbrains/kotlinx/dl/dataset/onHeapDatasetExtensions.kt
@@ -1,11 +1,11 @@
 /*
- * Copyright 2020 JetBrains s.r.o. and Kotlin Deep Learning project contributors. All Rights Reserved.
+ * Copyright 2020-2022 JetBrains s.r.o. and Kotlin Deep Learning project contributors. All Rights Reserved.
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
 package org.jetbrains.kotlinx.dl.dataset
 
-import org.jetbrains.kotlinx.dl.api.core.shape.shape
+import org.jetbrains.kotlinx.dl.api.core.shape.tensorShape
 import kotlin.math.roundToInt
 
 
@@ -23,7 +23,7 @@ public fun OnHeapDataset.fullToString(): String = buildStringRepr(x.contentDeepT
  * Builds intermediate [OnHeapDataset] string representation.
  */
 public fun OnHeapDataset.buildStringRepr(xString: String, yString: String): String =
-    "OnHeapDataset(\nx ${x.shape} =\n${xString},\ny [${y.size}] =\n${yString}\n)"
+    "OnHeapDataset(\nx ${x.tensorShape} =\n${xString},\ny [${y.size}] =\n${yString}\n)"
 
 
 /**
